% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-binBedpe.R
\name{binAnchor}
\alias{binAnchor}
\title{Define helper function for binning an anchor}
\usage{
binAnchor(a, p, res)
}
\arguments{
\item{a}{GRanges object}

\item{p}{Position within anchors to resize the bin. Can be a character or integer vector of length 1 or length(a) designating the position for each element in bedpe. Character options are "start", "end" and "center". Integers are referenced from the start position for '+' and '*' strands and from the end position for the '-' strand.}

\item{res}{Integer - resolution in which to bin the anchor.}
}
\value{
GRanges object that has been shifted and binned into res by p.
}
\description{
Define helper function for binning an anchor
}
\examples{

## Create example GRanges
gr1 <- GRanges(seqnames = "chr1",
               ranges = IRanges(start = rep(5000,3), end = rep(6000,3)),
               strand = c('+', '-', '*'))

gr2 <- gr1 \%>\% promoters(upstream = 2000, downstream = 200)

## Binning the results
binAnchor(gr1, 'start', 1000)
binAnchor(gr1, 'end', 1000)
binAnchor(gr1, 'center', 1000)

## Bin after shifting back to TSS
binAnchor(gr2, 2000, 1000)


}
