% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcApa.R
\name{calcApa}
\alias{calcApa}
\title{Extract BEDPE interactions from a Hi-C matrix}
\usage{
calcApa(
  bedpe,
  hicFile,
  norm = "NONE",
  res = 10000,
  buffer = 5,
  equalLengthOut = T
)
}
\arguments{
\item{bedpe}{data.frame or data.table in BEDPE format.}

\item{hicFile}{string path to .hic file.}

\item{norm}{string hic normalization <NONE/VC/VC_SQRT/KR>.}

\item{res}{integer resolution of bedpe bins.}

\item{buffer}{integer number of res-length bins from the center pixel.}

\item{equalLengthOut}{logical can be set to false to get an APA list that is not the
same length as the number of rows of the input bedpe file. Adds a slight efficieny
boost.}
}
\value{
Returns a list APA matricies that can be compiled and plotted.
}
\description{
Extract BEDPE interactions from a Hi-C matrix
}
