% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_binned}
\alias{check_binned}
\title{Helper function to check that a GInteractions object has been binned}
\usage{
check_binned(x, res)
}
\arguments{
\item{x}{GInteractions object.}

\item{res}{Integer - resolution in base-pairs.}
}
\description{
Helper function to check that a GInteractions object has been binned
}
