% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hictoolsr.R
\docType{package}
\name{hictoolsr-package}
\alias{hictoolsr}
\alias{hictoolsr-package}
\title{hictoolsr: a set of tools for Hi-C data analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

hictoolsr is an R package that contains a collection of useful tools for Hi-C data analysis.
}
\author{
\strong{Maintainer}: Eric Davis \email{esdavis@live.unc.edu}

}
