% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loopCounts}
\alias{loopCounts}
\title{Merged loop calls with contact frequency from replicate Hi-C experiments}
\format{
a GInteractions object with metadata columns describing
  loop properties and interaction counts from replicate
  Hi-C data
}
\usage{
data("loopCounts")
}
\description{
Using data from "Phase separation drives aberrant chromatin looping and
cancer development" by Ahn et al. 2021 (GEO links GSE143465 and GSE143465),
looping interactions were called with SIP (https://github.com/PouletAxel/SIP).
Resulting loop calls were merged using functions from `hictoolsr` and
interactions were extracted from replicate `.hic` files. The resulting file
can be used to demonstrate differential loop calling between two biological
conditions.
}
\references{
Ahn, Jeong Hyun, Eric S. Davis, Timothy A. Daugird,
 Shuai Zhao, Ivana Yoseli Quiroga, Hidetaka Uryu, Jie Li,
 et al. 2021. “Phase Separation Drives Aberrant Chromatin
 Looping and Cancer Development.” Nature 595 (7868): 591–95.

 SIP (https://github.com/PouletAxel/SIP)
}
\keyword{datasets}
