% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeBedpe.R
\name{mergeBedpe}
\alias{mergeBedpe}
\title{Read-in and merge BEDPE Interactions by Genomic distance with DBSCAN}
\usage{
mergeBedpe(
  bedpeFiles,
  res = 5000,
  selectCol = 12,
  dist_method = "manhattan",
  minPts = 2
)
}
\arguments{
\item{bedpeFiles}{Character vector of BEDPE file paths to be merged.}

\item{res}{Integer - resolution in base-pairs of all BEDPE interactions. Used to
determine the epsilon value for \code{dbscan()} (i.e. \code{eps = res*2}).}

\item{selectCol}{Integer or character string denoting the column to be used to select
clustered interactions.}

\item{dist_method}{Character string - distance measure used in the dist function. For
available methods see \code{?dist()}.}

\item{minPts}{Integer - minimum number of interactions to form a DBSCAN cluster.}
}
\value{
Returns a data table of selected BEDPE interactions
}
\description{
Given a list of BEDPE files, a resolution, and a selector column, this function will
use dbscan to merge loops by genomic distance.
}
