% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotApa.R
\name{plotApa}
\alias{plotApa}
\title{Plots Aggregate Peak Analysis Matrix}
\usage{
plotApa(
  params = NULL,
  apa,
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  just = c("left", "top"),
  default.units = "inches",
  draw = TRUE,
  palette = colorRampPalette(c("white", "dark red")),
  n = NULL,
  zrange = NULL
)
}
\arguments{
\item{params}{optional "pgParams" object containing relevant function parameters}

\item{x}{numeric or unit object specifying x-location of plot}

\item{y}{numeric or unit object specifying y-location of plot}

\item{width}{numeric or unit object specifying width of plot}

\item{height}{numeric or unit object specifying height of plot}

\item{just}{string or numeric vector specifying the justification of the viewport
relative to its (x, y) location}

\item{default.units}{string indicating the default units to use if x, y, width, or
height are only given as numeric vectors}

\item{draw}{logical value indicating whether graphics output should be produced}

\item{palette}{colorRampPalette function to use to map values to colors}

\item{n}{numeric by which to divide aggregate matrix}

\item{zrange}{vector of length 2; max and min values to set color scale}

\item{data}{matrix, list of matricies, or 3 column data.frame of APA results}
}
\value{
Function will draw a APA matrix and return an object of class "apa"
}
\description{
Plots Aggregate Peak Analysis Matrix
}
\examples{

## Create divergent matrix ####
m <- matrix(data = rnorm(n = 21*21, mean = 0, sd = 2), nrow = 21, ncol = 21)

## Define parameters
p <- pgParams(width = 3, height = 3, default.units = "inches")

## Create page
pageCreate(params = p)

## Plot apa
plot <- plotApa(apa = m,
                x = p$width/2, y = p$height/2,
                width = p$width*0.5, height = p$width*0.5, just = c("center", "center"),
                palette = colorRampPalette(c("blue", "white", "red")), zrange = NULL)

## Annotate legend
annoHeatmapLegend(plot = plot,
                  x = 2.3, y = 0.75, width = 0.1, height = 0.75)


## Create sequential matrix
m <- matrix(data = sample(0:100, 21*21, replace = T), nrow = 21, ncol = 21)

## Define parameters
p <- pgParams(width = 3, height = 3, default.units = "inches")

## Create page
pageCreate(params = p)

## Plot apa
plot <- plotApa(apa = m,
                x = p$width/2, y = p$height/2,
                width = p$width*0.5, height = p$width*0.5, just = c("center", "center"),
                palette = colorRampPalette(c("white", "dark red")), zrange = NULL)

## Annotate legend
annoHeatmapLegend(plot = plot,
                  x = 2.3, y = 0.75, width = 0.1, height = 0.75)


}
