% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcApa.R
\name{calcApa}
\alias{calcApa}
\title{Extract BEDPE interactions from a Hi-C matrix}
\usage{
calcApa(bedpe, hic, norm = "NONE", res = 10000, buffer = 5, filter = FALSE)
}
\arguments{
\item{bedpe}{data.frame or data.table in BEDPE format.}

\item{hic}{string path to .hic file.}

\item{norm}{string hic normalization <NONE/VC/VC_SQRT/KR>.}

\item{res}{integer resolution of bedpe bins.}

\item{buffer}{integer number of res-length bins from the center pixel.}

\item{filter}{TRUE or FALSE (default). If TRUE, \code{filterBedpe()} will be used to remove short intrachromosomal interactions that would otherwise cross the diagonal.}
}
\value{
Returns a list APA matricies that can be compiled and plotted.
}
\description{
Extract BEDPE interactions from a Hi-C matrix
}
